requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.10.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.9.0
torchvision==0.24.0
typing_extensions==4.15.0
Werkzeug==3.1.3
